Fix build on gcc and docs.

+++ src/engine_lib/src/render/texture_manager.c
@@ -93,7 +93,7 @@ texture_manager_request_texture(te_texture_manager* manager, const char* relativ
 
             glBindTexture(GL_TEXTURE_2D, tex_id);
             {
-                glTexImage2D(GL_TEXTURE_2D, 0, gl_internal_format, width, height, 0, gl_format,
+                glTexImage2D(GL_TEXTURE_2D, 0, gl_internal_format, width, height, 0, (unsigned int)gl_format,
                              GL_UNSIGNED_BYTE, pixels);
                 if (opt == TE_TLO_GENERATE_MIPMAPS) {
                     glGenerateMipmap(GL_TEXTURE_2D);
